REF: split the unit_test.go
