+++ ascii/test_struct.go
@@ -1,0 +1,7 @@
+package main
+
+type Args struct {
+	input        string
+	fileContents []string
+	asciiMap     map[rune]int
+}
